fix fix cyclic dependence again
